requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
backoff==2.2.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorlog==6.8.2
curl-cffi==0.6.2
datasets==2.14.7
dill==0.3.7
distro==1.9.0
dspy-ai==2.3.4
duckduckgo-search==4.5.0
filelock==3.13.1
frozenlist==1.4.1
fsspec==2023.10.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
huggingface-hub==0.21.3
idna==3.6
joblib==1.3.2
lxml==5.1.0
mako==1.3.2
markupsafe==2.1.5
multidict==6.0.5
multiprocess==0.70.15
numpy==1.26.4
openai==1.13.3
optuna==3.5.0
packaging==23.2
pandas==2.2.1
pyarrow==15.0.0
pyarrow-hotfix==0.6
pycparser==2.21
pydantic==2.5.0
pydantic-core==2.14.1
python-dateutil==2.9.0
pytz==2024.1
pyyaml==6.0.1
regex==2023.12.25
requests==2.31.0
six==1.16.0
sniffio==1.3.1
sqlalchemy==2.0.27
tqdm==4.66.2
typing-extensions==4.10.0
tzdata==2024.1
ujson==5.9.0
urllib3==2.2.1
xxhash==3.4.1
yarl==1.9.4
